+++ app/src/main/java/com/kinandcarta/create/proxytoggle/broadcast/ProxyUpdateListenerProviderImpl.kt
@@ -1,5 +1,6 @@
 package com.kinandcarta.create.proxytoggle.broadcast
 
+import com.kinandcarta.create.proxytoggle.feature.widget.broadcast.WidgetProxyUpdateListener
 import javax.inject.Inject
 
 class ProxyUpdateListenerProviderImpl @Inject constructor(
