+++ app/src/main/java/com/kinandcarta/create/proxytoggle/feature/manager/view/ProxyState.kt
@@ -1,4 +1,4 @@
-package com.kinandcarta.create.proxytoggle.view.manager
+package com.kinandcarta.create.proxytoggle.feature.manager.view
 
 sealed class ProxyState {
     data class Enabled(val address: String, val port: String) : ProxyState()
